requirements.txt
# CloudDrive2 Streaming Media Application

fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Media processing
ffmpeg-python==0.2.0

# Caching
aiocache==0.12.2
redis==5.0.1

# Utilities
httpx==0.25.2
